cloud/ipm/requirements.txt
# IPM 图像生成和斑马线检测依赖
# Python 3.12+

# ROS bag 文件读取
rosbags>=0.9.20

# 图像处理
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# 科学计算
scipy>=1.10.0

# 配置文件解析
PyYAML>=6.0

# 命令行工具
click>=8.0.0

# 表格显示
prettytable>=3.8.0

# 可选依赖（注释掉，按需安装）
# tqdm>=4.65.0    # 用于进度显示